# internal/store/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package db

import (
	uuid "github.com/gofrs/uuid/v5"
	"github.com/jackc/pgx/v5/pgtype"
)

type FibbingItAnswer struct {
	ID        uuid.UUID
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
	Answer    string
	PlayerID  uuid.UUID
	RoundID   uuid.UUID
	IsReady   pgtype.Bool
}

type FibbingItPlayerRole struct {
	ID         uuid.UUID
	CreatedAt  pgtype.Timestamp
	UpdatedAt  pgtype.Timestamp
	PlayerRole string
	RoundID    uuid.UUID
	PlayerID   uuid.UUID
}

type FibbingItRound struct {
	ID               uuid.UUID
	CreatedAt        pgtype.Timestamp
	UpdatedAt        pgtype.Timestamp
	RoundType        string
	Round            int32
	FibberQuestionID uuid.UUID
	NormalQuestionID uuid.UUID
	GameStateID      uuid.UUID
}

type FibbingItScore struct {
	ID        uuid.UUID
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
	PlayerID  uuid.UUID
	Score     int32
	RoundID   uuid.UUID
}

type FibbingItVote struct {
	ID               uuid.UUID
	CreatedAt        pgtype.Timestamp
	UpdatedAt        pgtype.Timestamp
	PlayerID         uuid.UUID
	VotedForPlayerID uuid.UUID
	RoundID          uuid.UUID
	IsReady          pgtype.Bool
}

type GameState struct {
	ID             uuid.UUID
	CreatedAt      pgtype.Timestamp
	UpdatedAt      pgtype.Timestamp
	RoomID         uuid.UUID
	SubmitDeadline pgtype.Timestamp
	State          string
}

type Player struct {
	ID        uuid.UUID
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
	Avatar    string
	Nickname  string
	IsReady   pgtype.Bool
	Locale    pgtype.Text
}

type Question struct {
	ID        uuid.UUID
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
	GameName  string
	RoundType string
	Enabled   pgtype.Bool
	GroupID   uuid.UUID
}

type QuestionsGroup struct {
	ID        uuid.UUID
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
	GroupName string
	GroupType string
}

type QuestionsI18n struct {
	ID         uuid.UUID
	CreatedAt  pgtype.Timestamp
	UpdatedAt  pgtype.Timestamp
	Question   string
	Locale     string
	QuestionID uuid.UUID
}

type Room struct {
	ID         uuid.UUID
	CreatedAt  pgtype.Timestamp
	UpdatedAt  pgtype.Timestamp
	GameName   string
	HostPlayer uuid.UUID
	RoomState  string
	RoomCode   string
}

type RoomsPlayer struct {
	RoomID    uuid.UUID
	PlayerID  uuid.UUID
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
}